+++ GEMValidation/interface/AnalyzerManager.h
@@ -2,7 +2,7 @@
 #define GEMCode_GEMValidation_AnalyzerManager_h
 
 #include "GEMCode/GEMValidation/interface/MatcherSuperManager.h"
-#include "GEMCode/GEMValidation/interface/TreeManager.h"
+#include "GEMCode/GEMValidation/interface/new/TreeManager.h"
 #include "GEMCode/GEMValidation/interface/Analyzers/GenParticleAnalyzer.h"
 #include "GEMCode/GEMValidation/interface/Analyzers/SimTrackAnalyzer.h"
 #include "GEMCode/GEMValidation/interface/Analyzers/GEMSimHitAnalyzer.h"
@@ -31,7 +31,7 @@ class AnalyzerManager
   void setMatcher(const MatcherManager&);
 
   /// do the matching
-  void analyze(const edm::Event& ev, const edm::EventSetup& es, const MatcherSuperManager& manager, TreeManager& tree);
+  void analyze(const edm::Event& ev, const edm::EventSetup& es, const MatcherSuperManager& manager, my::TreeManager& tree);
   void analyze(TreeManager& tree, const SimTrack& t, const SimVertex& v);
 
  private:
