+++ GEMValidation/interface/Analyzers/CSCStubAnalyzer.h
@@ -21,7 +21,7 @@ public:
   void setMatcher(const CSCStubMatcher& match_sh);
 
   // initialize the event
-  void analyze(const edm::Event& ev, const edm::EventSetup& es, const MatcherSuperManager& manager, TreeManager& tree);
+  void analyze(const edm::Event& ev, const edm::EventSetup& es, const MatcherSuperManager& manager, my::TreeManager& tree);
   void analyze(TreeManager& tree);
 
  private:
