+++ GEMValidation/interface/Analyzers/L1MuAnalyzer.h
@@ -18,7 +18,7 @@ public:
   void setMatcher(const L1MuMatcher& match_sh);
 
   // initialize the event
-  void analyze(const edm::Event& ev, const edm::EventSetup& es, const MatcherSuperManager& manager, TreeManager& tree);
+  void analyze(const edm::Event& ev, const edm::EventSetup& es, const MatcherSuperManager& manager, my::TreeManager& tree);
   void analyze(TreeManager& tree);
 
  private:
