+++ GEMValidation/interface/NewStructs/BaseStruct.h
@@ -1,0 +1,13 @@
+#ifndef GEMCode_GEMValidation_BaseStruct
+#define GEMCode_GEMValidation_BaseStruct
+
+#include "TTree.h"
+#include <vector>
+#include <string>
+
+typedef std::vector<float>* p_floats;
+typedef std::vector<int>* p_ints;
+typedef std::vector<float> t_floats;
+typedef std::vector<int> t_ints;
+
+#endif
