+++ GEMValidation/src/AnalyzerManager.cc
@@ -37,10 +37,9 @@ void AnalyzerManager::setMatcher(const MatcherManager& manager)
 void AnalyzerManager::analyze(const edm::Event& ev,
                               const edm::EventSetup& es,
                               const MatcherSuperManager& manager,
-                              TreeManager& tree) {
+                              my::TreeManager& tree) {
   simt_->analyze(ev, es, tree);
   gent_->analyze(ev, es, manager, tree);
-  /*
   gemsh_->analyze(ev, es, manager, tree);
   gemdg_->analyze(ev, es, manager, tree);
   gemstub_->analyze(ev, es, manager, tree);
@@ -49,7 +48,6 @@ void AnalyzerManager::analyze(const edm::Event& ev,
   cscstub_->analyze(ev, es, manager, tree);
   l1mu_->analyze(ev, es, manager, tree);
   l1track_->analyze(ev, es, manager, tree);
-  */
 }
